requirements.txt
streamlit==1.28.0
pdfplumber==0.7.6
python-docx==0.8.11
sentence-transformers==2.2.2
rapidfuzz==2.13.7
sqlalchemy==1.4.52
pandas==2.1.3
python-multipart==0.0.6
